new IpSpaceSpecifierFactory for constant IpWildcardSetIpSpaces
A new constant IpSpaceSpecifierFactory for ConstantIpSpaceSpecifiers
with IpWildcardSetIpSpaces. Uses a simple format for expressing
IpWildcardSetIpSpaces in a single string:
     W1,W2,... - W3,W4,...
or:  W1,W2,... \ W3,W4,...

etc, where the '-' or '\' separates the whitelist from the blacklist.

Also save the name of ConstantUniverseIpSpaceSpecifierFactory in a
constant field for testing, etc.

(cherry picked from commit 255d8bd)